Add files via upload
To run, save this to a folder and open a PowerShell instance. Run by typing the location such as "C:\Scripts\Project1.ps1". This script uses the Get-WmiObject cmdlet to retrieve information about logical disks on the system. It then iterates through each drive, checking if it's a valid drive (not a CD-ROM) and adds up the free space. Finally, it outputs the total number of drives and the total free space in gigabytes.